+++ Library/components/OpusBrand/index.ts
@@ -1,0 +1,2 @@
+export { OpusBrand } from "./OpusBrand";
+export type { OpusBrandProps, OpusBrandVariant } from "./OpusBrand";